wip: refactoring option parsing
