+++ starport-lambda/src/main/scala/com/krux/starport/lambda/SubmitHandler.scala
@@ -10,6 +10,21 @@ import org.slf4j.LoggerFactory
   * -Dstarport.config.url=s3://{your-bucket}/starport/starport.conf
   *
   */
+
+//sealed case class ExitException(status: Int) extends SecurityException("System.exit() called but ignored") {
+//}
+//
+//sealed class NoExitSecurityManager extends SecurityManager {
+//  override def checkPermission(perm: Permission): Unit = {}
+//
+//  override def checkPermission(perm: Permission, context: Object): Unit = {}
+//
+//  override def checkExit(status: Int): Unit = {
+//    super.checkExit(status)
+//    throw ExitException(status)
+//  }
+//}
+
 class SubmitHandler extends RequestHandler[SubmitRequest, SubmitResponse] {
 
   def logger = LoggerFactory.getLogger(getClass)
